Build 64 on hp-11
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@173730 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/util/HashPtr.cpp
@@ -67,7 +67,7 @@ HashPtr::~HashPtr()
 
 unsigned int HashPtr::getHashVal(const void *const key, unsigned int mod)
 {
- return ((long)key % mod);
+ return ((long)key % (unsigned long)mod);
 }
 
 bool HashPtr::equals(const void *const key1, const void *const key2)
